refactor: `Footer` and `GitHubStats` into separate files
